Create a materialised view for company_rankings
Companies are ranked on an outcome by looking at their outcome_values
for that outcome and whether higher_is_better.

For a group of outcomes, companies are ranked based on the average rank
over all outcomes in the group.

If a company has less than half the outcome_values for outcomes in a
group then this company does not receive a rank. It is set to NULL.

This is implemented with a materialised view that uses a recursive
query. We start at the bottom of the ‘tree’ with outcomes, calculate
company rankings and then move on to the level above.

Recursive queries don’t support aggregate functions so we use window
functions to calculate avg and count.

To nullify the rank of companies with missing outcome_values, we check
how many there are and nullify the rank if this is less than half the
size of the group.